Fixed ESP build issues
This mostly just puts all the String references into the m8r namespace to avoid conflicts with the Arduino version. It also fixed some signed/unsigned issues

+++ lib/m8rscript/src/ExecutionUnit.h
@@ -200,12 +200,12 @@ private:
     
     void checkEventQueueConsistency()
     {
-        int index = 0;
+        uint32_t index = 0;
         while (index < _eventQueue.size()) {
             assert(_eventQueue.size() - index >= 3);
             assert(_eventQueue[index + 2].isInteger());
             int32_t nargs = _eventQueue[index + 2].asIntValue();
-            assert(static_cast<int>(_eventQueue.size()) >= nargs + index + 3);
+            assert(static_cast<int32_t>(_eventQueue.size()) >= nargs + static_cast<int32_t>(index) + 3);
             index += nargs + 3;
         }
         assert(index == _eventQueue.size());
